pyproject.toml
[project]
name = "stock_predictor"
version = "0.1.0"
description = "stock_predictor using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
  "crewai>=0.51.0",
  "litellm>=1.42.0",
  "yfinance>=0.2.40",
  "scikit-learn>=1.4.0",
  "boto3>=1.34.0",
  "python-dotenv>=1.0.1",
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pandas>=2.2.0",
  "numpy>=1.26.0"
]

[project.scripts]
stock_predictor = "stock_predictor.main:run"
run_crew = "stock_predictor.main:run"
train = "stock_predictor.main:train"
replay = "stock_predictor.main:replay"
test = "stock_predictor.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
